telephony: Don't bypass telephony stack when switching RAT.
If we set the value for preferred network type prior to actually
receiving successful RAT change on other sim, then we may end up
in a scenario where the first network set completes sucessfully,
but the second (the one we set the value for) doesn't. If a user
reboots, they may then end up with a setting displaying a RAT
they didn't expect.

Ticket: CYNGNOS-674

Change-Id: Ib3ee9e46c70e6b98229a665f3ba5bc4f00baef1d